modfication version 33443

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -13,7 +13,8 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
+         if (!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table){
             $table->id();
             $table->string('name');
             $table->string('email')->unique();
@@ -23,6 +24,7 @@ return new class extends Migration
             $table->timestamps();
         });
     }
+    }
 
     /**
      * Reverse the migrations.
@@ -31,6 +33,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users');
+       
+        if (app()->environment('local')) {
+            Schema::dropIfExists('users');
+        }
     }
 };
